Added more pictures and objects. Got the labels to be the time of creation of the picture objects.
Issues:
copyImages.sh currently copies every png from a src to a dest (also the values need changed to work on local machine), rather than just checking to see if they are new (that's a to-do).
In createPictures.py it uses a counter for the id of the new images. However, this counter is local, so every time you create_pictures() it will override the old ones starting at zero (that's another to-do).